backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
redis[asyncio]==5.2.0
PyJWT==2.10.1
bcrypt==4.2.1
python-multipart==0.0.17
pydantic-settings==2.6.1
aiofiles==24.1.0
firebase-admin==6.6.0
python-dotenv==1.0.1
alembic==1.14.0
pytest-html==4.1.1
psycopg2-binary==2.9.10

# LLM support
anthropic==0.84.0      # Anthropic Claude API + Bedrock
openai==2.28.0         # Ollama (OpenAI-compatible function calling)
